+++ web/static/js/containers/QuarantineIdeasListContainer.jsx
@@ -4,13 +4,41 @@ import { connect } from 'react-redux';
 import QuarantineIdeasListItem from '../components/QuarantineIdeasListItem';
 
 import IdeasList from '../components/IdeasList';
+import IdeasFilter from '../components/IdeasFilter';
+import { changeQuarantineFilterStatus, changeQuarantineFilterText } from '../actions/ideas';
 
-const QuarantineIdeasListContainer = ({ ideas }) => (
-  <IdeasList ideas={ideas.map(idea => (<QuarantineIdeasListItem key={idea.id} idea={idea} />))} />
+const visibleIdeas = (ideas, filter) => {
+  let filteredIdeas = [...ideas].filter(i => (i.status === 'under_review' || i.status === 'denied'));
+  if (filter.text.length > 0) {
+    filteredIdeas = filteredIdeas
+      .filter(i => i.title.toUpperCase().indexOf(filter.text.toUpperCase()) !== -1);
+  }
+
+  return filteredIdeas.filter(i => filter.status.includes(i.status));
+};
+
+const QuarantineIdeasListContainer = ({ ideas, filter, onChangeFilterText, onChangeFilterStatus }) => ( // eslint-disable-line max-len
+  <div>
+    <IdeasFilter
+      filter={filter}
+      onChange={onChangeFilterText}
+      onClickStatus={onChangeFilterStatus}
+      statuses={['under_review', 'denied']}
+    />
+    <IdeasList ideas={ideas.map(idea => (<QuarantineIdeasListItem key={idea.id} idea={idea} />))} />
+  </div>
 );
 
 QuarantineIdeasListContainer.propTypes = {
-  ideas: PropTypes.array,
+  ideas: PropTypes.arrayOf(PropTypes.shape({
+    status: PropTypes.string.isRequired,
+  })),
+  filter: PropTypes.shape({
+    status: PropTypes.arrayOf(PropTypes.string).isRequired,
+    text: PropTypes.string.isRequired,
+  }).isRequired,
+  onChangeFilterText: PropTypes.func.isRequired,
+  onChangeFilterStatus: PropTypes.func.isRequired,
 };
 
 QuarantineIdeasListContainer.defaultProps = {
@@ -18,7 +46,17 @@ QuarantineIdeasListContainer.defaultProps = {
 };
 
 const mapStateToProps = state => ({
-  ideas: state.ideas.filter(i => (i.status === 'under_review')),
+  ideas: visibleIdeas(state.ideas, state.quarantineIdeaFilter),
+  filter: state.quarantineIdeaFilter,
 });
 
-export default connect(mapStateToProps)(QuarantineIdeasListContainer);
+const mapDispatchToProps = dispatch => ({
+  onChangeFilterText: (event) => {
+    dispatch(changeQuarantineFilterText(event.target.value));
+  },
+  onChangeFilterStatus: status => () => {
+    dispatch(changeQuarantineFilterStatus(status));
+  },
+});
+
+export default connect(mapStateToProps, mapDispatchToProps)(QuarantineIdeasListContainer);