+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Use a better color for under revision ideas
 - Fix bug causing users to always be set to kind 'user' on login
+- Allow user to filter quarantined ideas, and show under_revision ideas
 
 # 3.0.3